// Web/RMA-Web-Shell/projects/fincare/src/app/payment-manager/models/commission-account.ts
export class CommissionAccount {
    accountTypeId: number;
    accountTypeName: string;
    accountId: number;
    accountCode: string;
    accountName: string;
    identificationNumber: string;
    totalPendingRelease: number;
    totalSubmitted: number;
    totalWithHeld: number;
    totalPaid: number;
    totalRejected: number;
    clawBackAccountBalance: number;
}