// Web/RMA-Web-Shell/projects/fincare/src/app/payment-manager/models/commission-pool-search-params.ts
export class CommissionPoolSearchParams {
    pageIndex: number;
    pageSize: number;
    orderBy: string;
    direction: string;
    currentQuery: string;
    reAllocate: boolean;
    userLoggedIn: number;
    workPoolId: number;
    startDate: string;
    endDate: string;
    commissionStatusId: number;
}